Ensure dir exists before rmdir()

+++ src/Atlas/Dir/Local.php
@@ -486,7 +486,9 @@ class Local implements
             }
         }
 
-        rmdir($this->path);
+        if ($this->exists()) {
+            rmdir($this->path);
+        }
     }
 
     /**
